asobi-be/asobi-auth-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
google-cloud-firestore==2.16.1
google-cloud-storage==2.18.2
google-cloud-logging==3.11.3
httpx==0.27.2
pytest==8.3.3
email-validator>=2.1.0
python-dotenv>=1.0.1
pydantic-settings>=2.4.0
orjson>=3.9.0